+++ data_provider/__init__.py
@@ -36,7 +36,17 @@ from .tushare_fetcher import TushareFetcher
 from .pytdx_fetcher import PytdxFetcher
 from .baostock_fetcher import BaostockFetcher
 from .yfinance_fetcher import YfinanceFetcher
-from .us_index_mapping import is_us_index_code, is_us_stock_code, get_us_index_yf_symbol, US_INDEX_MAPPING
+from .us_index_mapping import (
+    is_us_index_code,
+    is_us_stock_code,
+    get_us_index_yf_symbol,
+    US_INDEX_MAPPING,
+    is_jp_stock_code,
+    is_jp_index_code,
+    get_jp_index_yf_symbol,
+    JP_INDEX_MAPPING,
+    get_jp_stock_yf_symbol,
+)
 
 __all__ = [
     'BaseFetcher',
@@ -52,4 +62,9 @@ __all__ = [
     'is_hk_stock_code',
     'get_us_index_yf_symbol',
     'US_INDEX_MAPPING',
+    'is_jp_stock_code',
+    'is_jp_index_code',
+    'get_jp_index_yf_symbol',
+    'JP_INDEX_MAPPING',
+    'get_jp_stock_yf_symbol',
 ]